Add more unit tests.
also rename the tests so they can be run using:

python -m unittest discover
