+++ tests/testgetoptions.py
